+++ protected/controllers/AdminManuscriptController.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 class AdminManuscriptController extends Controller
 {
 
@@ -49,18 +51,27 @@ class AdminManuscriptController extends Controller
 	 */
 	public function actionCreate()
 	{
-		$model=new Manuscript;
+		$model = new \GigaDB\models\Manuscript();
+        $modelWrapper = new LegacyManuscriptForm($model);
 
-		if ($attrs = Yii::$app->request->post('Manuscript'))
-		{
+		if ($attrs = Yii::$app->request->post('LegacyManuscriptForm')) {
+            //otherwise problem with id not null constraint
+            $model = new \GigaDB\models\Manuscript();
 			$model->attributes = $attrs;
 
-			if ($model->save())
-				return $this->redirect(array('view','id'=>$model->id));
+			if ($model->save()) {
+                return $this->redirect(array('view','id'=>$model->id));
+            } else {
+                foreach ($model->getErrors() as $attribute => $errors) {
+                    foreach ($errors as $error) {
+                        $modelWrapper->addError($attribute, $error);
+                    }
+                }
+            }
 		}
 
 		$this->render('create',array(
-			'model'=>$model,
+			'model'=>$modelWrapper,
 		));
 	}
 
@@ -71,9 +82,10 @@ class AdminManuscriptController extends Controller
 	 */
 	public function actionUpdate($id)
 	{
-		$model=$this->loadModel($id);
+		$model=$this->loadYii2Model($id);
+        $modelWrapper = new LegacyManuscriptForm($model);
 
-        if ($attrs = Yii::$app->request->post('Manuscript'))
+        if ($attrs = Yii::$app->request->post('LegacyManuscriptForm'))
 		{
 			$model->attributes = $attrs;
 			if ($model->save())
@@ -81,7 +93,7 @@ class AdminManuscriptController extends Controller
 		}
 
 		$this->render('update',array(
-			'model'=>$model,
+			'model'=>$modelWrapper,
 		));
 	}
 
@@ -144,6 +156,17 @@ class AdminManuscriptController extends Controller
 			throw new CHttpException(404,'The requested page does not exist.');
 		return $model;
 	}
+
+    public function loadYii2Model($id): \GigaDB\models\Manuscript
+    {
+        $model = \GigaDB\models\Manuscript::findOne($id);
+
+        if (!$model) {
+            throw new \yii\web\NotFoundHttpException('The requested page does not exist.');
+        }
+
+        return $model;
+    }
 
 	/**
 	 * Performs the AJAX validation.