+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Feat #2235: Create Manuscript and DatasetLog Yii2 models and create a wrapper to use Manuscript Yii2 model with yii1 Form
 - Feat #2235: Add a checkbox to determine whether if a manuscript should be considered a pre-print link
 - Fix #2294: Add dedicated username variable for authenticating to email service